Backend/requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
cvzone
h5py
pillow
